move playlist card into cards
